Ironed out a lotta stuff with pointers
pointers can now be used as params, basically as variables everywhere. it took waaaay too long to work it out tho.

IMP TODO: The processtuple thing has a MAJOR bug, the tuple items are stored in a ll, and then the entire ll is set as children for a tuple node. but the items' parent is never set to this tuple node. FIX IT ASAP.